requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-filter==2.4.0
django-rest-swagger==2.2.0
django-sendsms==0.5
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.27.2
drf-user==1.1.0
drf-yasg==1.21.7
drfaddons==0.1.4
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.1
openapi-codec==1.3.2
packaging==24.1
pillow==10.4.0
PyJWT==1.7.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
